+++ src/d3-funnel/D3Funnel.js
@@ -329,6 +329,7 @@ class D3Funnel {
 
             return {
                 index,
+                data: rawBlock,
                 ratio,
                 value: block.value,
                 height: this.settings.height * ratio,